==> notebooks/5_llm_processing.ipynb <==
import sys
import os
import pandas as pd
import json
from dotenv import load_dotenv
import tiktoken
import pickle

from langchain_core.documents import Document
from langchain.text_splitter import RecursiveCharacterTextSplitter
from langchain_core.rate_limiters import InMemoryRateLimiter

from langchain.vectorstores import FAISS
from langchain.prompts import PromptTemplate
from langchain.chains import RetrievalQA

from langchain.agents import AgentExecutor, Tool, initialize_agent, load_tools
import json


module_path = os.path.abspath(os.path.join('..'))
if module_path not in sys.path:
    sys.path.append(module_path)

from src.llm import get_azure_embeddings_client, get_llm_client, get_gemini_llm_client

if not load_dotenv():
    raise Exception('Error loading .env file. Make sure to place valid keys in the .env file.')

ARTICLES_CLEAN_DIR = os.path.join("..", "data", "articles_clean")
FILTERED_METADATA_PATH = os.path.join("..", "data", "filtered_metadata.csv")
DB_PATH = os.path.join("..", "data", "db", "sample.db")
RESULTS_DIR =  os.path.join("..", "data", "results")

if not os.path.exists(DB_PATH):
    os.makedirs(os.path.dirname(DB_PATH), exist_ok=True)

if not os.path.exists(RESULTS_DIR):
    os.makedirs(os.path.dirname(RESULTS_DIR), exist_ok=True)

# Default go-to Openrouter LLM - check README for other available models
llm = get_llm_client(
    # Configurable parameters
    max_tokens=1024,
    temperature=0.2,
)

system_prompt = """
You are an expert assistant. Use only the provided news document to create a list of the main entities in the document.
Assign each person a relevance score from 0 to 1 based on their impact in the news article. For example, the main subject 
should have a score of 1, and someone who only appears once to contribute a quote should receive a score near 0. Each entity
should also have an entity_type attribute that describes the type of entity (it should be person if the entity is a human)
Furthermore, create a list of relationships between entity in the story. Each relationship should have a descriptor (for example 
"empolyee of", "enemy of", "brother"), and an attitude attribute (1 for strong allies, -1 for committed enemies, 0 for ambivalent). 
If the nature of the relationship cannot be determined, use 0 as the attitude, and leave the descriptor blank. There should be a 
relationship between each pair of entity in the first list. 
The output should be structured as a JSON file with two fields:
entities: the list of entity, each an object with name and relevance
relationships: the list of relationships, each an object with originator, target, descriptor, and attitude.
Double check that the output is valid JSON.
Document: {context}
"""

prompt_template = PromptTemplate(
    input_variables=["context"], 
    template=system_prompt
)

# Define a tool to validate JSON
def validate_json(output):
    """Checks if output is valid JSON."""
    try:
        json.loads(output)
        return True
    except json.JSONDecodeError:
        return False

# Define the agent
def create_json_agent():
    agent = initialize_agent(
        tools=[],  # No extra tools needed
        llm=llm,
        agent="zero-shot-react-description",
        verbose=True
    )
    return agent

# Create the agent
agent = create_json_agent()

# Define a function to run the query and ensure valid JSON output
def query_with_json_validation(query):
    valid_output = False
    attempts = 0
    max_attempts = 5
    result = ""

    while not valid_output and attempts < max_attempts:
        attempts += 1
        print(f"Attempt {attempts}...")
        response = agent.run(prompt_template.format(query=query))
        if validate_json(response):
            valid_output = True
            result = response
        else:
            print("Invalid JSON. Retrying...")

    if not valid_output:
        raise ValueError("Failed to get valid JSON after multiple attempts.")

    return result

filtered_metadata = pd.read_csv(FILTERED_METADATA_PATH)
filtered_metadata.head(5)

def get_documents_from_path(filenames: list[str]) -> list[Document]:
    documents = []
    
    for file_name in filenames:
        file_path = os.path.join(ARTICLES_CLEAN_DIR, file_name)
        with open(file_path, "r", encoding="utf-8") as file:
            file = json.load(file)

        text = file.get("text", "")
        documents.append(Document(page_content=text, metadata={
            "title": file.get("title", ""),
            "author": file.get("author", ""),
            "published_at": file.get("published_at", ""),
            "id": file.get("id", ""),
        }))

    return documents

documents = get_documents_from_path(filtered_metadata["filename"])
print(f"Number of articles: {len(documents)}")

jsons = []
for document in documents:
    result_filename = f'{document.metadata['id']}.json'
    result_filepath = os.path.join(RESULTS_DIR, result_filename)
    print(result_filepath)
    if os.path.exists(result_filepath):
        print('Skipping')
        continue
    query_input = document.page_content
    result: str = query_with_json_validation(query_input)
    jsons.append(result)
    cleaned = result.removeprefix("```json").removesuffix('```')
    try:
        result_json = json.JSONDecoder().decode(cleaned)
        with open(result_filepath, "w") as file:
            json.dump(result_json, file, indent=4)
    except Exception:
        print('Error')
        continue
